refactor(metrics): drop passthrough barrel files
Every module lived in its own directory beside an index.ts that
re-exported the single identically named sibling. Neither the directory
nor the barrel earned anything: tsdown bundles from src/index.ts, so the
barrels never reached dist, and every spec already imported the
implementation file directly.

Move the implementations up a level and delete the 14 barrels. Most
import specifiers are untouched, since '../client' and './client'
resolved to the same module before and after.

timing.ts is the one behavioural edit. Its barrel converted a default
export into the named `timing` the package exports publicly, so the
default becomes a named export instead. The default was never public.

environment/ and normalizers/ keep their index.ts. Both group more than
one implementation and do more than re-export a single sibling.
