--- packages/metrics/src/client/index.ts
@@ -1,3 +1,0 @@
-import { configure, defaultRegister, Prometheus } from './client';
-
-export { Prometheus, defaultRegister, configure };
