--- packages/metrics/src/create-gc-metrics/index.ts
@@ -1,3 +1,0 @@
-import { createGcMetrics } from './create-gc-metrics';
-
-export { createGcMetrics };
