--- packages/metrics/src/create-gc-observer/index.ts
@@ -1,3 +1,0 @@
-import { createGcObserver } from './create-gc-observer';
-
-export { createGcObserver };
