--- packages/metrics/src/create-graphql-metrics/index.ts
@@ -1,3 +1,0 @@
-import { createGraphQlMetrics } from './create-graphql-metrics';
-
-export { createGraphQlMetrics };
