+++ packages/metrics/src/create-http-metrics.spec.js
@@ -1,4 +1,4 @@
-vi.mock('../client', () => ({
+vi.mock('./client', () => ({
   configure: vi.fn(),
   defaultRegister: {
     getSingleMetric: vi.fn(),
