--- packages/metrics/src/create-http-metrics/index.ts
@@ -1,3 +1,0 @@
-import { createHttpMetrics } from './create-http-metrics';
-
-export { createHttpMetrics };
