--- packages/metrics/src/create-metric/index.ts
@@ -1,8 +1,0 @@
-import {
-  createCounter,
-  createGauge,
-  createHistogram,
-  createSummary,
-} from './create-metric';
-
-export { createCounter, createGauge, createHistogram, createSummary };
