+++ packages/metrics/src/create-request-recorder.ts
@@ -6,10 +6,10 @@ import type {
 } from '@promster/types';
 import merge from 'merge-options';
 
-import { endMeasurementFrom } from '../end-measurement-from';
-import { skipMetricsInEnvironment } from '../environment';
-import { sortLabels } from '../sort-labels';
-import type { Timing } from '../timing';
+import { endMeasurementFrom } from './end-measurement-from';
+import { skipMetricsInEnvironment } from './environment';
+import { sortLabels } from './sort-labels';
+import type { Timing } from './timing';
 
 type TRecordingOptions = {
   labels: TLabelValues;
