--- packages/metrics/src/create-request-recorder/index.ts
@@ -1,4 +1,0 @@
-import { createRequestRecorder } from './create-request-recorder';
-
-export type { TRequestRecorder } from './create-request-recorder';
-export { createRequestRecorder };
