+++ packages/metrics/src/end-measurement-from.spec.js
