+++ packages/metrics/src/end-measurement-from.ts
