--- packages/metrics/src/end-measurement-from/index.ts
@@ -1,3 +1,0 @@
-import { endMeasurementFrom } from './end-measurement-from';
-
-export { endMeasurementFrom };
