+++ packages/metrics/src/normalizers/method.spec.js
