--- packages/metrics/src/normalizers/method/index.ts
@@ -1,3 +1,0 @@
-import { normalizeMethod } from './method';
-
-export { normalizeMethod };
