--- packages/metrics/src/normalizers/path/index.ts
@@ -1,3 +1,0 @@
-import { normalizePath } from './path';
-
-export { normalizePath };
