--- packages/metrics/src/normalizers/status-code/index.ts
@@ -1,3 +1,0 @@
-import { normalizeStatusCode } from './status-code';
-
-export { normalizeStatusCode };
