+++ packages/metrics/src/sort-labels.spec.js
