--- packages/metrics/src/sort-labels/index.ts
@@ -1,3 +1,0 @@
-import { sortLabels } from './sort-labels';
-
-export { sortLabels };
