+++ packages/metrics/src/summary.ts
@@ -1,4 +1,4 @@
-import { defaultRegister } from '../client';
+import { defaultRegister } from './client';
 
 const getSummary = async () => defaultRegister.metrics();
 const getContentType = () => defaultRegister.contentType;
