--- packages/metrics/src/summary/index.ts
@@ -1,3 +1,0 @@
-import { getContentType, getSummary } from './summary';
-
-export { getSummary, getContentType };
