--- packages/metrics/src/timing/index.ts
@@ -1,3 +1,0 @@
-import timing, { Timing } from './timing';
-
-export { timing, Timing };
